+++ spark/src/main/scala/org/apache/comet/serde/operator/CometGlobalLimit.scala
@@ -17,12 +17,13 @@
  * under the License.
  */
 
-package org.apache.comet.serde
+package org.apache.comet.serde.operator
 
 import org.apache.spark.sql.execution.GlobalLimitExec
 
 import org.apache.comet.{CometConf, ConfigEntry}
 import org.apache.comet.CometSparkSessionExtensions.withInfo
+import org.apache.comet.serde.{CometOperatorSerde, OperatorOuterClass}
 import org.apache.comet.serde.OperatorOuterClass.Operator
 
 object CometGlobalLimit extends CometOperatorSerde[GlobalLimitExec] {
